Graf Algorithms | Breadth-First Search Shortest Path with Python
Breadth first search is one of the basic and essential searching algorithms on graphs. As a result of how the algorithm works, the path found by breadth first search to any node is the shortest path to that node, i.e the path that contains the smallest number of edges in unweighted graphs.

Co-Authored-By: Faru Nuri Sönmez <23061257+FaruNuriSonmez@users.noreply.github.com>